+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed
+
+* Added `update_type` to `OrderUpdate` and `TradeUpdate` (#39)
+
 ## 0.8.2 &ndash; 2022-02-18
 
 ### Changed